VMManager: Change block offsets to size_t

+++ src/core/hle/kernel/vm_manager.cpp
@@ -55,7 +55,7 @@ VMManager::VMAHandle VMManager::FindVMA(VAddr target) const {
 }
 
 ResultVal<VMManager::VMAHandle> VMManager::MapMemoryBlock(VAddr target,
-        std::shared_ptr<std::vector<u8>> block, u32 offset, u32 size, MemoryState state) {
+        std::shared_ptr<std::vector<u8>> block, size_t offset, u32 size, MemoryState state) {
     ASSERT(block != nullptr);
     ASSERT(offset + size <= block->size());
